+++ lib/cider.dart
@@ -1,4 +1,5 @@
 library cider;
 
-export 'package:cider/src/cider.dart';
-export 'package:cider/src/cider_command.dart';
+export 'package:cider/src/project.dart';
+export 'package:cider/src/cli/cider_cli.dart';
+export 'package:cider/src/cli/console.dart';