+++ lib/src/changelog_file.dart
@@ -1,0 +1,6 @@
+import 'dart:io';
+
+import 'package:path/path.dart';
+
+File changelogFile(String projectRoot) =>
+    File(join(projectRoot, 'CHANGELOG.md'));