+++ lib/src/pubspec_file.dart
@@ -1,0 +1,5 @@
+import 'dart:io';
+
+import 'package:path/path.dart';
+
+File pubspecFile(String projectRoot) => File(join(projectRoot, 'pubspec.yaml'));