+++ test/functional_test.dart
@@ -1,46 +1,75 @@
 import 'dart:io';
 
-import 'package:cider/cider.dart';
-import 'package:path/path.dart';
+import 'package:cider/src/cli/channel.dart';
+import 'package:cider/src/cli/cider_cli.dart';
+import 'package:cider/src/cli/console.dart';
+import 'package:path/path.dart' as path;
+import 'package:pubspec_parse/pubspec_parse.dart';
 import 'package:test/test.dart';
-
-import 'src/testing.dart';
 
 void main() {
   late Directory temp;
-  late Cider cider;
-  late MockStdout out, err;
+  final out = BufferChannel();
+  final err = BufferChannel();
+  final cli = CiderCli(console: Console(out: out, err: err));
+
+  Future<int> run(List<String> args) =>
+      cli.run(['--project-root=${temp.absolute.path}', ...args]);
+
   setUp(() async {
+    out.buffer.clear();
+    err.buffer.clear();
     temp = await Directory.systemTemp.createTemp();
-    await Directory('test/template').list().forEach((element) {
+    await Directory('test/template').list().forEach((element) async {
       if (element is File) {
-        element.copy(join(temp.path, basename(element.path)));
+        await element.copy(path.join(temp.path, path.basename(element.path)));
       }
     });
-    out = MockStdout();
-    err = MockStdout();
-    cider = Cider(root: temp);
-    cider.provide<Stdout>((_) => out);
-    cider.provide<Stdout>((_) => err, name: 'stderr');
   });
 
   tearDown(() async {
     await temp.delete(recursive: true);
   });
 
+  group('Smoke', () {
+    test('Can read the project version', () async {
+      final actual = (await Process.run(
+              Platform.executable, ['bin/cider.dart', 'version']))
+          .stdout as String;
+      final expected = Pubspec.parse(await File('pubspec.yaml').readAsString())
+          .version!
+          .toString();
+      expect(actual.trim(), equals(expected));
+    });
+  });
+
+  test('Usage exception', () async {
+    final code = await run(['foo']);
+    expect(code, equals(64));
+    expect(err.buffer.toString().trim(), contains('Available commands:'));
+  });
+
+  test('Can read the project version', () async {
+    await cli.run(['version']);
+    final expected = Pubspec.parse(await File('pubspec.yaml').readAsString())
+        .version!
+        .toString();
+    expect(out.buffer.toString().trim(), equals(expected));
+  });
+
   test('Full release cycle', () async {
-    final code = await cider.run(['log', 'add', 'Initial release']);
+    final code = await run(['log', 'add', 'Initial release']);
     expect(code, 0);
-    await cider.run(['describe']);
+    await run(['describe']);
     final step1 = '''
 ## Unreleased
 ### Added
 - Initial release
 ''';
     expect(out.buffer.toString(), step1);
-    await cider.run(['version', '1.0.0']);
+    await run(['version', '1.0.0']);
     out.buffer.clear();
-    await cider.run(['release', '--date=2020-01-02']);
+    await run(['release', '--date=2020-01-02']);
     final step2 = '''
 ## [1.0.0] - 2020-01-02
 ### Added
@@ -49,24 +78,37 @@ void main() {
 [1.0.0]: https://github.com/example/project/releases/tag/1.0.0
 ''';
     expect(out.buffer.toString(), step2);
-    out.buffer.clear();
-    await cider.run(['log', 'change', 'New turbo V6 engine installed']);
-    await cider.run(['log', 'fix', 'Wheels falling off sporadically']);
-    await cider.run(['describe']);
-    final step3 = '''
-## [Unreleased]
+    await run(['log', 'change', 'New turbo V6 engine installed']);
+    await run(['log', 'fix', 'Wheels falling off sporadically']);
+    final step3Body = '''
 ### Changed
 - New turbo V6 engine installed
 
 ### Fixed
 - Wheels falling off sporadically
-
+''';
+    final step3 = '''
+## [Unreleased]
+$step3Body
 [Unreleased]: https://github.com/example/project/compare/1.0.0...HEAD
 ''';
-    expect(out.buffer.toString(), step3);
-    await cider.run(['bump', 'minor']);
     out.buffer.clear();
-    await cider.run(['release', '--date=2021-02-03']);
+    await run(['describe']);
+    expect(out.buffer.toString(), step3);
+
+    // TODO: implement in 0.2.1
+    // out.buffer.clear();
+    // await run(['describe', '-b']);
+    // expect(out.buffer.toString(), step3Body);
+
+    await run(['bump', 'minor']);
+
+    out.buffer.clear();
+    await run(['list', '-y', '-u']);
+    expect(out.buffer.toString(), 'Unreleased\n1.0.0\n');
+
+    out.buffer.clear();
+    await run(['release', '--date=2021-02-03']);
     final step4 = '''
 ## [1.1.0] - 2021-02-03
 ### Changed
@@ -79,7 +121,7 @@ void main() {
 ''';
     expect(out.buffer.toString(), step4);
     out.buffer.clear();
-    await cider.run(['yank', '1.1.0']);
+    await run(['yank', '1.1.0']);
     final step5 = '''
 ## [1.1.0] - 2021-02-03 \\[YANKED\\]
 ### Changed
@@ -91,25 +133,41 @@ void main() {
 [1.1.0]: https://github.com/example/project/compare/1.0.0...1.1.0
 ''';
     expect(out.buffer.toString(), step5);
+
     out.buffer.clear();
-    await cider.run(['unyank', '1.1.0']);
+    await run(['list']);
+    expect(out.buffer.toString(), '1.0.0\n');
+
+    out.buffer.clear();
+    await run(['list', '-y']);
+    expect(out.buffer.toString(), '1.1.0\n1.0.0\n');
+
+    out.buffer.clear();
+    await run(['list', '-y', '-u']);
+    expect(out.buffer.toString(), '1.1.0\n1.0.0\n');
+
+    out.buffer.clear();
+    await run(['unyank', '1.1.0']);
     expect(out.buffer.toString(), step4);
+    expect(err.buffer.toString(), isEmpty);
   });
 
   group('Version', () {
     test('get', () async {
-      final code = await cider.run(['version']);
+      final code = await run(['version']);
       expect(code, 0);
       expect(out.buffer.toString().trim(), '0.0.5-alpha+42');
+      expect(err.buffer.toString(), isEmpty);
     });
 
     test('set', () async {
-      final code = await cider.run(['version', '1.0.0']);
+      final code = await run(['version', '1.0.0']);
       expect(code, 0);
       expect(out.buffer.toString().trim(), '1.0.0');
       out.buffer.clear();
-      await cider.run(['version']);
+      await run(['version']);
       expect(out.buffer.toString().trim(), '1.0.0');
+      expect(err.buffer.toString(), isEmpty);
     });
 
     group('bump', () {
@@ -146,35 +204,54 @@ void main() {
         ['bump', 'release', '--keep-build']: '0.0.5+42',
       }.forEach((args, expected) {
         test('${args.join(' ')} => $expected', () async {
-          final code = await cider.run(args);
+          final code = await run(args);
           expect(code, 0);
+          expect(err.buffer.toString(), isEmpty);
           out.buffer.clear();
-          await cider.run(['version']);
+          await run(['version']);
           expect(out.buffer.toString().trim(), expected);
         });
+      });
+      test('version must increase', () async {
+        final code = await run(['bump', 'build', '--keep-build']);
+        expect(code, 65);
+        expect(err.buffer.toString().trim(),
+            'The next version must be higher than the current one.');
+      });
+      test('version part must be specified', () async {
+        final code = await run(['bump']);
+        expect(code, 65);
+        expect(err.buffer.toString().trim(), 'Version part must be specified');
       });
     });
   });
 
   group('Find Root', () {
     test('valid root', () async {
-      final code = await cider.run(['version']);
-      expect(code, Cider.exitOK);
+      final code = await cli.run(['version']);
+      expect(code, 0);
     });
 
     test('invalid root', () async {
-      Directory fsRoot = Directory.current;
+      Directory systemRoot = Directory.current;
 
-      while (!FileSystemEntity.identicalSync(fsRoot.path, fsRoot.parent.path)) {
-        fsRoot = fsRoot.parent;
+      while (!FileSystemEntity.identicalSync(
+          systemRoot.path, systemRoot.parent.path)) {
+        systemRoot = systemRoot.parent;
       }
-
-      Cider ciderAtRoot = Cider(root: fsRoot);
-      ciderAtRoot.provide<Stdout>((_) => out);
-      ciderAtRoot.provide<Stdout>((_) => err, name: 'stderr');
-      final code = await ciderAtRoot.run(['version']);
-      expect(code, Cider.exitException);
-      expect(err.buffer.toString(), contains("Can not find project root"));
+      final code = await runIn(systemRoot, () => cli.run(['version']));
+      expect(code, 70);
+      expect(err.buffer.toString(), contains('Can not find project root'));
     });
   });
+}
+
+Future<T> runIn<T>(Directory dir, Future<T> Function() f) async {
+  final current = Directory.current;
+  Directory.current = dir;
+  try {
+    return await f();
+  } finally {
+    Directory.current = current;
+  }
 }