+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2023-06-10
+### Added
+- The `list` command to list all versions from the changelog.
+- Restored the `--project-root` argument.
+
+### Changed
+- Bump min SDK version to 3.0.0.
+- Exit codes to closer match POSIX.
+
 ## [0.1.6] - 2023-06-07
 ### Fixed
 - Correctly abort upwards search for project root directory upon reaching filesystem root. (thanks [@maltevesper](https://github.com/maltevesper))
@@ -82,6 +91,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial version
 
+[0.2.0]: https://github.com/f3ath/cider/compare/0.1.6...0.2.0
 [0.1.6]: https://github.com/f3ath/cider/compare/0.1.5...0.1.6
 [0.1.5]: https://github.com/f3ath/cider/compare/0.1.4...0.1.5
 [0.1.4]: https://github.com/f3ath/cider/compare/0.1.3...0.1.4